Feature/server/project file refactoring (#152)
* rename execution web

* move register admin functions to admin module

* add setup to execution module

* move register scenario endpoints to scenario module

* remove web module, move login to admin, register endpoints in app, add global error handler

* remove deprecated import

* fix import

* fix value error on setup

* pr change request

* move request blocker, add test

* make test-flask-app global for test

* remove lasting app creation

* fix circular import

---------
