+++ server/scenario/web_api/setup.py
@@ -1,8 +1,8 @@
 from flask import Flask
 
-from execution.web.notification import web_api
+from scenario.web_api import scenario
 
 
 def setup(app: Flask):
     """ Connects each implemented endpoint to flask environment. """
-    app.register_blueprint(web_api, url_prefix="/web")
+    app.register_blueprint(scenario.web_api, url_prefix="/web")
