--- server/web/__init__.py
@@ -1,1 +1,0 @@
-from .setup import setup
