--- server/web/api/test_data.py
@@ -1,1 +1,0 @@
-test_execution = {}
